main/requirements.txt
beautifulsoup4==4.11.1
chromadb==0.5.12
fastapi==0.115.2
googlesearch_python==1.2.5
ollama==0.3.3
# psycopg2==2.9.9
psycopg2_binary==2.9.9
pydantic==2.9.2
Requests==2.32.3
transformers==4.41.0
uvicorn==0.31.0
nltk==3.7
langchain==0.3.6
langchain-core==0.3.15
langchain-community==0.3.4
streamlit==1.29.0
sentence_transformers==3.2.1
plotly
pandas
numpy
seaborn
duckduckgo-search==6.3.3
prometheus-client
prometheus-fastapi-instrumentator
opik
Werkzeug